JCF: Issue #361: very rough draft of a fddatautilities build; works locally on daq.fnal.gov
fddatautilities is an umbrella package which depends on a couple of
far-detector-focused packages as well as the core set of packages
(currently named "dunedaq" as it has for a couple of years, but soon
to be renamed "coredaq"). dunedaq has been given a build variant
called "subset" which when it assumes the value of "datautilities"
will only depend on the subset of core packages which fddatautilities
needs.

Note that it may make sense to have the variant value match up
perfectly with the name of the target package it's intended to
support.

In order to not try and solve everything at once, the package.py files
for the umbrella packages needed for this build are located in a
temporary subdirectory called issue361_dev_files. With this
subdirectory I can sidestep worrying about the make-release-repo.py
algorithm whenever I want to modify and test the relevant package.py
files for the umbrella packages during the development process.

I've successfully build fddatautilities using the files in this commit
via calling build-release.sh manually inside of a container on daq.fnal.gov
which is volume-mounted to a v2.1 externals area.