JCF: Issue #500: switch to a new directory structure to accommodate variants of coredaq + add YAML files for fddatautilities
The idea here is, a typical target package won't need to use *all* the packages
in coredaq (the canonical exception currently being fddaq, of course). To account
for this, the coredaq packages a target package needs will be listed in
"configs/coredaq/PACKAGE_NAME/release.yaml". This means, e.g., we rename
configs/coredaq/coredaq-develop/release.yaml to configs/coredaq/fddaq-develop/release.yaml
to reflect the fact that the original coredaq can now be considered the fddaq
variant of coredaq.

I've also added a package to aid in my development and testing, called "fddatautilities".
This is de-facto an alias for rawdatautils, but it means I can try using an fddatautilities
environment to run file_quality_checker.py, as a crosscheck for this framework.