Compress nightly workflow schedule
This is motivated by a recently-noticed issue in which the CI Dashboard
workflow was not publishing "extended" integration test results because
the workflow times were overlapping. I'm not exactly sure how this is
happening, since they both run on the same self-hosted runner and the
latter job should wait in the queue until the former is complete. But I
can see in the timestamps of the GitHub Actions logs that it does
happen.